pyproject.toml
[project]
name = "deepresearchagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.22.0",
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.0.1",
    "google-genai>=0.8.5",
    "google-generativeai>=0.8.5",
]
